requirements.txt
aiofiles==24.1.0
aioredis==2.0.1
alembic==1.16.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
azure-cognitiveservices-language-textanalytics==0.2.2
azure-common==1.1.28
azure-core==1.34.0
azure-identity==1.23.0
azure-mgmt-core==1.5.0
brotli==1.1.0
certifi==2025.6.15
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
cryptography==45.0.4
cssselect2==0.8.0
defusedxml==0.7.1
distro==1.9.0
fastapi==0.115.13
fonttools==4.58.4
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
isodate==0.7.2
jiter==0.10.0
mako==1.3.10
markdown==3.8
markupsafe==3.0.2
msal==1.32.3
msal-extensions==1.3.1
msrest==0.7.1
oauthlib==3.3.0
openai==1.88.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.9.1
pydyf==0.11.0
pyjwt==2.10.1
pyodbc==5.2.0
pyphen==0.17.2
python-dotenv==1.1.0
python-multipart==0.0.20
pytube==15.0.0
pyyaml==6.0.2
redis==6.2.0
reportlab==4.4.1
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.41
starlette==0.46.2
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.1.0
weasyprint==65.1
webencodings==0.5.1
websockets==15.0.1
youtube-transcript-api==1.1.0
zopfli==0.2.3.post1

python-pptx
pypdf
python-docx
